Add support for multipart title naming templates
